libril: Do not enforce checks on data registration state response length
libril expects the response length of the result of DATA_REGISTRATION_STATE vendor call as either 6*sizeof(char*) or 11*sizeof(char*). If the the response length is not one of them, it treats it as an invalid response. In case of taoshan, its vendor RIL returns 12(*sizeof(char*)) as result of this call, which makes libril treat it as invalid respone, thus causing data registration failure.

Thus, make this check permissive, and make libril do not treat the response as invalid in such cases. Also, such checks where not done in Nougat libril.
This change fixes data registration and hence mobile data.

Signed-off-by: Avinaba Dalal <d97.avinaba@gmail.com>